Remove certain images and organize them into folders
