+++ frontend/front/src/pages/Public/Layout/Navbar.js
@@ -23,7 +23,7 @@ import {
 import { styled } from "@mui/material/styles";
 import CloseIcon from "@mui/icons-material/Close";
 import ButtonGetPump from "../Components/Button/ButtonGetPump";
-import logoHeatPump from "../../../assets/images/bhpa-logo.gif";
+import logoHeatPump from "../../../assets/images/bhpa-logos/bhpa-logo300px.gif";
 import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
 import ExpandLess from "@mui/icons-material/ExpandLess";
 import ExpandMore from "@mui/icons-material/ExpandMore";
